refactor(client): decuppling git and github
