+++ packages/amplication-client/src/Application/github/dialogs/GitDialogsContainer.tsx
@@ -53,6 +53,7 @@ export default function GitDialogsContainer({
         <GithubRepos
           applicationId={app.id}
           onCompleted={handleSelectRepoDialogDismiss}
+          sourceControlService={sourceControlService}
         />
       </Dialog>
       <Dialog
